Refactored sourc code, use consistent style
So far, the style of the source code was rather inconsistent, mixing
space and tab indentation, different formatting styles and bracket
conventions. Using clang, the code was automatically formatted such that
a consistent style is used across all files. A variation of the llvm
style was applied. For indentation, tabs with width 4 are used. The following
command was executed for an automatic conversion:

    $ clang-format -style="{BasedOnStyle: llvm, IndentWidth: 4, \
        UseTab: ForIndentation, TabWidth: 4}" -i *.{cc,hh}

Furthermore, the .gitignore now also includes .root-files generated by
the simulation and temporary files generated by editors.