Added 2013 setup for 164Dy (high energy) (run263–run283)
The setup for the 2013 experimental campaign measuring high energy E1
strength in 164Dy was added. This configuration is relevant for run263
to run283. The level of detail this simulation is able to provide
surpasses the amount of information that was documented, so several
assumptions were made based on the recent 2016 compaign.

The 2nd setup was removed.. The detectors were moved to their assumed
positions, filters and both targets (Dy164 and Dy164_2O3) were added,
including target containers.