ParticleSD position and momentum recording
- Previously, they were obtained by invoking G4Track::GetPosition(), which
gives the position/momentum after a step. Now,
G4Step::GetPreStepPoint()->GetPosition() is used, which gives the
correct result.
- Noticed this bug when trying to record particles only when they are
detected at specific volume boundaries. Before this fix, the hit point
which came shortly after the volume crossing was recorded, and not the
hit point on the boundary. A simulation with geantinos uncovered the
problem.