Remove EMPhysicsPolarizedJAEA model
This model was seldomly used.
It is a modified version of G4EmLivermorePolarizedPhysics.
With Geant4 11.0, the original model was modified in a
backwards-incompatible way.
Keeping this modified version in the repository would require providing
different versions, depending on the Geant4 version.
This already caused problems with older Geant4 versions.
After discussion, we decided to remove it completely for easier
maintenance.