Increased clover crystal anode hole size
By comparing the simulated detector responses to 12C@15.11MeV-Circ-Pol
and 28Si@11.22MeV-Circ-Pol NRF lines with experimental data from the
2021 (154Sm) GDR experiment, it was found, that the simulated intensity
of SEP and DEP in comparison to the FEP of the clover detectors was too
small. Hence, the size of the anode hole was increased here to improve
the reproduction of the observed relative intensities of SEP and DEP in
the simulation.